WIP test reorg
